pyproject.toml
[project]
name = "rizzoclaw"
version = "0.1.0"
description = "Telegram bot that forwards messages to an AI agent"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.6",
    "aiosqlite>=0.22.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "sqlalchemy>=2.0.48",
    "pypdf",
    "reportlab"
]

[project.scripts]
rizzoclaw = "rizzoclaw.bot:main"
